+++ lib/multiple_search_selection.dart
@@ -35,26 +35,6 @@ class MultipleSearchSelection<T> extends StatefulWidget {
     List<T>? initialPickedItems,
     Widget? title,
     BoxDecoration? searchFieldBoxDecoration,
-    @Deprecated(
-      'Use pickedItemsBoxDecoration instead. This will be removed in the next major version.',
-    )
-        Color? pickedItemsBorderColor,
-    @Deprecated(
-      'Use searchFieldBoxDecoration instead. This will be removed in the next major version.',
-    )
-        Color? searchFieldBorderColor,
-    @Deprecated(
-      'Use searchFieldBoxDecoration instead. This will be removed in the next major version.',
-    )
-        Color? searchFieldBackgroundColor,
-    @Deprecated(
-      'Use showedItemsBoxDecoration instead. This will be removed in the next major version.',
-    )
-        Color? showedItemsBackgroundColor,
-    @Deprecated(
-      'Use showedItemsBoxDecoration instead. This will be removed in the next major version.',
-    )
-        Color? showedItemsBorderColor,
     double? showedItemsScrollbarMinThumbLength,
     Color? showedItemsScrollbarColor,
     double? showedItemsScrollbarMinOverscrollLength,
@@ -94,14 +74,16 @@ class MultipleSearchSelection<T> extends StatefulWidget {
     Widget? clearAllButton,
     VoidCallback? onTapClearAll,
     bool? caseSensitiveSearch,
-    TextEditingController? textEditingController,
+    TextEditingController? searchFieldTextEditingController,
     FocusNode? textFieldFocus,
     String hintText = 'Type here to search',
     double? showedItemExtent,
+    int? maxSelectedItems,
   }) =>
       MultipleSearchSelection._(
         items: items,
         title: title,
+        maxSelectedItems: maxSelectedItems,
         isCreatable: false,
         key: key ?? ValueKey(items.hashCode),
         clearSearchFieldOnSelect: clearSearchFieldOnSelect ?? false,
@@ -122,9 +104,6 @@ class MultipleSearchSelection<T> extends StatefulWidget {
         onTapShowItems: onTapShowItems,
         onTapShowedItem: onTapShowedItem,
         searchFieldBoxDecoration: searchFieldBoxDecoration,
-        searchFieldBackgroundColor: searchFieldBackgroundColor,
-        searchFieldBorderColor: searchFieldBorderColor,
-        pickedItemsBorderColor: pickedItemsBorderColor,
         pickedItemSpacing: pickedItemSpacing,
         pickedItemsBoxDecoration: pickedItemsBoxDecoration,
         pickedItemsContainerMaxHeight: pickedItemsContainerMaxHeight,
@@ -149,8 +128,6 @@ class MultipleSearchSelection<T> extends StatefulWidget {
         showShowedItemsScrollbar: showShowedItemsScrollbar,
         showedItemContainerHeight: showedItemContainerHeight,
         showedItemContainerPadding: showedItemContainerPadding,
-        showedItemsBackgroundColor: showedItemsBackgroundColor,
-        showedItemsBorderColor: showedItemsBorderColor,
         showedItemsBoxDecoration: showedItemsBoxDecoration,
         showedItemsScrollController:
             showedItemsScrollController ?? ScrollController(),
@@ -165,7 +142,7 @@ class MultipleSearchSelection<T> extends StatefulWidget {
         caseSensitiveSearch: caseSensitiveSearch ?? false,
         pickedItemsContainerBuilder: pickedItemsContainerBuilder,
         searchFieldTextEditingController:
-            textEditingController ?? TextEditingController(),
+            searchFieldTextEditingController ?? TextEditingController(),
         textFieldFocus: textFieldFocus ?? FocusNode(),
         hintText: hintText,
         showedItemExtent: showedItemExtent,
@@ -191,26 +168,6 @@ class MultipleSearchSelection<T> extends StatefulWidget {
     List<T>? initialPickedItems,
     Widget? title,
     BoxDecoration? searchFieldBoxDecoration,
-    @Deprecated(
-      'Use pickedItemsBoxDecoration instead. This will be removed in the next major version.',
-    )
-        Color? pickedItemsBorderColor,
-    @Deprecated(
-      'Use searchFieldBoxDecoration instead. This will be removed in the next major version.',
-    )
-        Color? searchFieldBorderColor,
-    @Deprecated(
-      'Use searchFieldBoxDecoration instead. This will be removed in the next major version.',
-    )
-        Color? searchFieldBackgroundColor,
-    @Deprecated(
-      'Use showedItemsBoxDecoration instead. This will be removed in the next major version.',
-    )
-        Color? showedItemsBackgroundColor,
-    @Deprecated(
-      'Use showedItemsBoxDecoration instead. This will be removed in the next major version.',
-    )
-        Color? showedItemsBorderColor,
     Color? showedItemsScrollbarColor,
     double? showedItemsScrollbarMinThumbLength,
     double? showedItemsScrollbarMinOverscrollLength,
@@ -253,10 +210,12 @@ class MultipleSearchSelection<T> extends StatefulWidget {
     FocusNode? textFieldFocus,
     String hintText = 'Type here to search',
     double? showedItemExtent,
+    int? maxSelectedItems,
   }) =>
       MultipleSearchSelection._(
         items: items,
         title: title,
+        maxSelectedItems: maxSelectedItems,
         isCreatable: true,
         createOptions: createOptions,
         key: key ?? ValueKey(items.hashCode),
@@ -276,11 +235,8 @@ class MultipleSearchSelection<T> extends StatefulWidget {
         onTapSelectAll: onTapSelectAll,
         onTapShowItems: onTapShowItems,
         onTapShowedItem: onTapShowedItem,
-        searchFieldBackgroundColor: searchFieldBackgroundColor,
-        searchFieldBorderColor: searchFieldBorderColor,
         searchFieldBoxDecoration: searchFieldBoxDecoration,
         pickedItemSpacing: pickedItemSpacing,
-        pickedItemsBorderColor: pickedItemsBorderColor,
         pickedItemsBoxDecoration: pickedItemsBoxDecoration,
         pickedItemsContainerMaxHeight: pickedItemsContainerMaxHeight,
         pickedItemsContainerMinHeight: pickedItemsContainerMinHeight,
@@ -304,7 +260,6 @@ class MultipleSearchSelection<T> extends StatefulWidget {
         showShowedItemsScrollbar: showShowedItemsScrollbar,
         showedItemContainerHeight: showedItemContainerHeight,
         showedItemContainerPadding: showedItemContainerPadding,
-        showedItemsBackgroundColor: showedItemsBackgroundColor,
         showedItemsBoxDecoration: showedItemsBoxDecoration,
         showedItemsScrollController:
             showedItemsScrollController ?? ScrollController(),
@@ -314,7 +269,6 @@ class MultipleSearchSelection<T> extends StatefulWidget {
             showedItemsScrollbarMinOverscrollLength,
         showedItemsScrollbarMinThumbLength: showedItemsScrollbarMinThumbLength,
         showedItemsScrollbarRadius: showedItemsScrollbarRadius,
-        showedItemsBorderColor: showedItemsBorderColor,
         sortPickedItems: sortPickedItems ?? false,
         sortShowedItems: sortShowedItems ?? false,
         caseSensitiveSearch: caseSensitiveSearch ?? false,
@@ -336,6 +290,7 @@ class MultipleSearchSelection<T> extends StatefulWidget {
     required this.searchFieldTextEditingController,
     required this.textFieldFocus,
     super.key,
+    this.maxSelectedItems,
     this.onPickedChange,
     this.createOptions,
     this.items,
@@ -359,32 +314,12 @@ class MultipleSearchSelection<T> extends StatefulWidget {
     this.showedItemsScrollbarRadius,
     this.showedItemContainerHeight,
     this.showedItemContainerPadding,
-    @Deprecated(
-      'Use showedItemsBoxDecoration instead. This will be removed in the next major version.',
-    )
-        this.showedItemsBackgroundColor,
-    @Deprecated(
-      'Use showedItemsBoxDecoration instead. This will be removed in the next major version.',
-    )
-        this.showedItemsBorderColor,
     this.showedItemsScrollPhysics,
     this.showedItemsBoxDecoration,
     this.searchFieldInputDecoration,
     this.noResultsWidget,
-    @Deprecated(
-      'Use outerContainerDecoration instead. This will be removed in the next major version.',
-    )
-        this.searchFieldBackgroundColor,
-    @Deprecated(
-      'Use outerContainerDecoration instead. This will be removed in the next major version.',
-    )
-        this.searchFieldBorderColor,
     this.searchFieldBoxDecoration,
     this.itemsVisibility = ShowedItemsVisibility.alwaysOn,
-    @Deprecated(
-      'Use pickedItemsBoxDecoration instead. This will be removed in the next major version.',
-    )
-        this.pickedItemsBorderColor,
     this.pickedItemSpacing,
     this.pickedItemsContainerMaxHeight,
     this.pickedItemsContainerMinHeight,
@@ -409,29 +344,20 @@ class MultipleSearchSelection<T> extends StatefulWidget {
     this.showedItemExtent,
   });
 
+  /// The maximum number of items that can be picked. If null, there is no limit.
+  ///
+  /// If the max number is met:
+  ///
+  /// - The [showItemsButton] will be disabled (if the item visibility is toggle).
+  /// - The search field will be disabled.
+  /// - The showed items will collapse.
+  final int? maxSelectedItems;
+
   /// The title widget on top of the picked items.
   final Widget? title;
 
-  /// The border color of the picked items container.
-  @Deprecated(
-    'Use pickedItemsBoxDecoration instead. This will be removed in the next major version.',
-  )
-  final Color? pickedItemsBorderColor;
-
   /// The [BoxDecoration] of the picked items container.
   final BoxDecoration? pickedItemsBoxDecoration;
-
-  /// The color of the container that includes the search text field and the showed items.
-  @Deprecated(
-    'Use outerContainerDecoration instead. This will be removed in the next major version.',
-  )
-  final Color? searchFieldBackgroundColor;
-
-  /// The border color of the container that includes the search text field and the showed items.
-  @Deprecated(
-    'Use outerContainerDecoration instead. This will be removed in the next major version.',
-  )
-  final Color? searchFieldBorderColor;
 
   /// The box decoration of the container that includes the search text field and the showed items.
   ///
@@ -452,18 +378,6 @@ class MultipleSearchSelection<T> extends StatefulWidget {
 
   /// The [BoxDecoration] of the showed items container.
   final BoxDecoration? showedItemsBoxDecoration;
-
-  /// The background color of the container holding all the items.
-  @Deprecated(
-    'Use showedItemsBoxDecoration instead. This will be removed in the next major version.',
-  )
-  final Color? showedItemsBackgroundColor;
-
-  /// The border color of the showed items container.
-  @Deprecated(
-    'Use showedItemsBoxDecoration instead. This will be removed in the next major version.',
-  )
-  final Color? showedItemsBorderColor;
 
   /// The minimum length of the items' scrollbar thumb.
   final double? showedItemsScrollbarMinThumbLength;
@@ -786,6 +700,12 @@ class _MultipleSearchSelectionState<T>
   }
 
   void _onAddItem(T item) {
+    if (widget.maxSelectedItems != null &&
+        pickedItems.length >= widget.maxSelectedItems!) {
+      Navigator.pop(context);
+      return;
+    }
+
     widget.onTapShowedItem?.call();
     final T pickedItem = item;
     pickedItems.add(pickedItem);
@@ -802,6 +722,14 @@ class _MultipleSearchSelectionState<T>
       pickedItems,
     );
     widget.onItemAdded?.call(pickedItem);
+
+    if (widget.maxSelectedItems != null &&
+        pickedItems.length == widget.maxSelectedItems!) {
+      widget.searchFieldTextEditingController.clear();
+      if (widget.itemsVisibility == ShowedItemsVisibility.toggle) {
+        Navigator.pop(context);
+      }
+    }
   }
 
   void _onCreateItem() {
@@ -830,7 +758,9 @@ class _MultipleSearchSelectionState<T>
   }
 
   void _selectAllItems() {
-    pickedItems.addAll(showedItems);
+    pickedItems.addAll(
+      showedItems,
+    );
     if (widget.sortPickedItems) {
       pickedItems.sort(
         (a, b) => widget.fieldToCheck(a).compareTo(
@@ -886,7 +816,7 @@ class _MultipleSearchSelectionState<T>
       itemCount: showedItems.isEmpty ? 1 : showedItems.length,
       itemExtent: widget.showedItemExtent,
       itemBuilder: (context, index) {
-        if (showedItems.isEmpty) {
+        if (showedItems.isEmpty && allItems.isNotEmpty) {
           return widget.isCreatable
               ? GestureDetector(
                   onTap: _onCreateItem,
@@ -933,6 +863,8 @@ class _MultipleSearchSelectionState<T>
 
   @override
   Widget build(BuildContext context) {
+    final bool maxItemsSelected = widget.maxSelectedItems != null &&
+        (widget.maxSelectedItems!) <= pickedItems.length;
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
@@ -964,8 +896,7 @@ class _MultipleSearchSelectionState<T>
                 BoxDecoration(
                   border: pickedItems.isNotEmpty
                       ? Border.all(
-                          color: widget.pickedItemsBorderColor ??
-                              Colors.grey.withOpacity(0.5),
+                          color: Colors.grey.withOpacity(0.5),
                         )
                       : null,
                 ),
@@ -1034,21 +965,19 @@ class _MultipleSearchSelectionState<T>
                                   mainAxisSize: MainAxisSize.min,
                                   children: [
                                     DecoratedBox(
-                                      decoration: widget
-                                              .searchFieldBoxDecoration ??
-                                          BoxDecoration(
-                                            color: widget
-                                                    .searchFieldBackgroundColor ??
-                                                Colors.white,
-                                            border: Border.all(
-                                              color: widget
-                                                      .searchFieldBorderColor ??
-                                                  Colors.grey.withOpacity(0.5),
-                                            ),
-                                          ),
+                                      decoration:
+                                          widget.searchFieldBoxDecoration ??
+                                              BoxDecoration(
+                                                color: Colors.white,
+                                                border: Border.all(
+                                                  color: Colors.grey
+                                                      .withOpacity(0.5),
+                                                ),
+                                              ),
                                       child: TextField(
                                         key: const Key('toggle-searchfield'),
                                         focusNode: widget.textFieldFocus,
+                                        enabled: !maxItemsSelected,
                                         controller: widget
                                             .searchFieldTextEditingController,
                                         style: widget.searchFieldTextStyle,
@@ -1095,27 +1024,19 @@ class _MultipleSearchSelectionState<T>
                                       decoration: widget
                                               .showedItemsBoxDecoration ??
                                           BoxDecoration(
-                                            color: widget
-                                                    .showedItemsBackgroundColor ??
-                                                Colors.grey.withOpacity(0.1),
+                                            color: Colors.grey.withOpacity(0.1),
                                             border: Border(
                                               bottom: BorderSide(
-                                                color: widget
-                                                        .showedItemsBorderColor ??
-                                                    Colors.grey
-                                                        .withOpacity(0.5),
+                                                color: Colors.grey
+                                                    .withOpacity(0.5),
                                               ),
                                               left: BorderSide(
-                                                color: widget
-                                                        .showedItemsBorderColor ??
-                                                    Colors.grey
-                                                        .withOpacity(0.5),
+                                                color: Colors.grey
+                                                    .withOpacity(0.5),
                                               ),
                                               right: BorderSide(
-                                                color: widget
-                                                        .showedItemsBorderColor ??
-                                                    Colors.grey
-                                                        .withOpacity(0.5),
+                                                color: Colors.grey
+                                                    .withOpacity(0.5),
                                               ),
                                             ),
                                           ),
@@ -1125,7 +1046,9 @@ class _MultipleSearchSelectionState<T>
                                                 .copyWith(
                                           scrollbars: false,
                                         ),
-                                        child: _buildShowedItems(),
+                                        child: maxItemsSelected
+                                            ? const SizedBox()
+                                            : _buildShowedItems(),
                                       ),
                                     )
                                   ],
@@ -1139,23 +1062,26 @@ class _MultipleSearchSelectionState<T>
                         });
                       },
                       child: IgnorePointer(
-                        child:
-                            widget.showItemsButton ?? const Text('Show items'),
+                        child: maxItemsSelected
+                            ? const SizedBox()
+                            : widget.showItemsButton ??
+                                const Text('Show items'),
                       ),
                     ),
                     const SizedBox(
                       width: 10,
                     ),
                   ],
-                  if (widget.showSelectAllButton ?? true)
-                    GestureDetector(
-                      behavior: HitTestBehavior.opaque,
-                      onTap: _selectAllItems,
-                      child: IgnorePointer(
-                        child:
-                            widget.selectAllButton ?? const Text('Select all'),
+                  if (widget.maxSelectedItems == null)
+                    if (widget.showSelectAllButton ?? true)
+                      GestureDetector(
+                        behavior: HitTestBehavior.opaque,
+                        onTap: _selectAllItems,
+                        child: IgnorePointer(
+                          child: widget.selectAllButton ??
+                              const Text('Select all'),
+                        ),
                       ),
-                    ),
                 ],
               ),
               if (pickedItems.isNotEmpty && (widget.showClearAllButton ?? true))
@@ -1176,29 +1102,26 @@ class _MultipleSearchSelectionState<T>
           DecoratedBox(
             decoration: widget.searchFieldBoxDecoration ??
                 BoxDecoration(
-                  color: widget.searchFieldBackgroundColor ?? Colors.white,
+                  color: Colors.white,
                   border: Border(
                     top: BorderSide(
-                      color: widget.searchFieldBorderColor ??
-                          Colors.grey.withOpacity(0.5),
+                      color: Colors.grey.withOpacity(0.5),
                     ),
                     left: BorderSide(
-                      color: widget.searchFieldBorderColor ??
-                          Colors.grey.withOpacity(0.5),
+                      color: Colors.grey.withOpacity(0.5),
                     ),
                     right: BorderSide(
-                      color: widget.searchFieldBorderColor ??
-                          Colors.grey.withOpacity(0.5),
+                      color: Colors.grey.withOpacity(0.5),
                     ),
                     bottom: BorderSide(
-                      color: widget.searchFieldBorderColor ??
-                          Colors.grey.withOpacity(0.5),
+                      color: Colors.grey.withOpacity(0.5),
                     ),
                   ),
                 ),
             child: TextField(
               key: const Key('searchfield'),
               focusNode: widget.textFieldFocus,
+              enabled: !maxItemsSelected,
               controller: widget.searchFieldTextEditingController,
               style: widget.searchFieldTextStyle,
               decoration: widget.searchFieldInputDecoration ??
@@ -1238,20 +1161,16 @@ class _MultipleSearchSelectionState<T>
             ),
             decoration: widget.showedItemsBoxDecoration ??
                 BoxDecoration(
-                  color: widget.showedItemsBackgroundColor ??
-                      Colors.grey.withOpacity(0.1),
+                  color: Colors.grey.withOpacity(0.1),
                   border: Border(
                     bottom: BorderSide(
-                      color: widget.showedItemsBorderColor ??
-                          Colors.grey.withOpacity(0.5),
+                      color: Colors.grey.withOpacity(0.5),
                     ),
                     left: BorderSide(
-                      color: widget.showedItemsBorderColor ??
-                          Colors.grey.withOpacity(0.5),
+                      color: Colors.grey.withOpacity(0.5),
                     ),
                     right: BorderSide(
-                      color: widget.showedItemsBorderColor ??
-                          Colors.grey.withOpacity(0.5),
+                      color: Colors.grey.withOpacity(0.5),
                     ),
                   ),
                 ),
@@ -1268,7 +1187,8 @@ class _MultipleSearchSelectionState<T>
               child: ScrollConfiguration(
                 behavior:
                     ScrollConfiguration.of(context).copyWith(scrollbars: false),
-                child: _buildShowedItems(),
+                child:
+                    maxItemsSelected ? const SizedBox() : _buildShowedItems(),
               ),
             ),
           )